+++ library/src/main/java/com/aspsine/multithreaddownload/entity/ThreadInfo.java
@@ -6,15 +6,15 @@ package com.aspsine.multithreaddownload.entity;
 public class ThreadInfo {
     private int id;
     private String url;
-    private int start;
-    private int end;
-    private int finished;
+    private long start;
+    private long end;
+    private long finished;
     private int progress;
 
     public ThreadInfo() {
     }
 
-    public ThreadInfo(int id, String url, int start, int end, int finished) {
+    public ThreadInfo(int id, String url, long start, long end, long finished) {
         this.id = id;
         this.url = url;
         this.start = start;
@@ -44,27 +44,27 @@ public class ThreadInfo {
         this.url = url;
     }
 
-    public int getStart() {
+    public long getStart() {
         return start;
     }
 
-    public void setStart(int start) {
+    public void setStart(long start) {
         this.start = start;
     }
 
-    public int getEnd() {
+    public long getEnd() {
         return end;
     }
 
-    public void setEnd(int end) {
+    public void setEnd(long end) {
         this.end = end;
     }
 
-    public int getFinished() {
+    public long getFinished() {
         return finished;
     }
 
-    public void setFinished(int finished) {
+    public void setFinished(long finished) {
         this.finished = finished;
     }
